+++ vorx/cli/commands/__init__.py
@@ -4,3 +4,4 @@ from . import list
 list.list['help']['function'] = main.help
 list.list['version']['function'] = main.version
 list.list['ping']['function'] = main.ping
+list.list['init']['function'] = main.init