+++ vorx/conf.py
@@ -1,4 +1,4 @@
-VERSION = "2026.0.1.1"
+VERSION = "2026.0.2"
 BUILD = "dev"
-BUILD_DATE = "2026-05-09"
+BUILD_DATE = "2026-05-12"
 SERVER_SOCK = "vorx.codcatdev.site"