+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # All Change Log history
 
+## [2026.0.2-dev] - 2026-05-12
+
+A "Core" update
+
+### Added
+- New CLI command - `init`
+- Started working on the main core of the engine
+- Added dependencies and setup instructions for the engine to ReadMe
+- Added `Vector2` class, with functions `length`, `normalized`, `normalize` (vorx/core/maths/vectors.pyx)
+- Added `build.py` script for building the engine Cython (.pyx) components to C. For more speed and better performance.
+
+### Fixed
+- Fixed a bug in the CLI with ping (On Server-Side)
+
+<hr>
+
 ## [2026.0.1.1-dev] - 2026-05-09
 
 A mini patch update